Boston Housing Price Prediction
Here, Boston1.csv file is used as the dataset to predict the Housing prices at Boston.
Programming Language used is: - Python (Jupyter Notebook)
Model used is Linear Regression (skLearn -> Supervised Learning Algorithm) :-
            1. Simple Linear Regression 
            2. Multi-Linear Regression 
After doing EDA, datset is seperated in training and testing data and then the model is fitted on the training dataset. Predictions are then 
made using the test data. Accuracy of the model is calculated based on the predictions.